refactor config directory search to use `$XDG_CONFIG_DIRS`
Replaces the use of `dirs::config_dir` with manual handling of the
`XDG_CONFIG_DIRS` environment variable to search for the configuration
directory We now check all directories listed in `XDG_CONFIG_DIRS` for
the existence of the `hyprlauncher` config folder. If no directory is
found, it falls back to the default value of `~/.config/hyprlauncher`.
Not ideal, but should still leeway for config handling